Add move for connection.
